Please read https://github.com/fontforge/fontforge/issues/401
We need to force configure.ac NOT to use libuninameslist if we find it.
We want to run the libunicodenames choice.
We won't modify m4/fontforge_create_pkg-config_files.m4 since we do not
want to call this a "dependency", but still choose this as an "option".
Do not apply this patch to mainline code.